requirements.txt
pandas
numpy
tensorflow
keras
pillow
scipy
matplotlib
scikit-learn
seaborn
albumentations 
opencv-python
customtkinter